refactor: BookedBlock FreeLayer components are moved to separated files.
